+++ lib/aixm/feature/navigational_aid/designated_point.rb
@@ -43,15 +43,22 @@ module AIXM
         end
 
         ##
-        # Render AIXM
+        # Render UID markup
+        def to_uid(*extensions)
+          builder = Builder::XmlMarkup.new(indent: 2)
+          builder.DpnUid({ newEntity: (true if extensions >> :ofm) }.compact) do |dpnuid|
+            dpnuid.codeId(id)
+            dpnuid.geoLat(xy.lat(format_for(*extensions)))
+            dpnuid.geoLong(xy.long(format_for(*extensions)))
+          end
+        end
+
+        ##
+        # Render AIXM markup
         def to_xml(*extensions)
           builder = to_builder(*extensions)
           builder.Dpn do |dpn|
-            dpn.DpnUid({ newEntity: (true if extensions >> :ofm) }.compact) do |dpnuid|
-              dpnuid.codeId(id)
-              dpnuid.geoLat(xy.lat(format_for(*extensions)))
-              dpnuid.geoLong(xy.long(format_for(*extensions)))
-            end
+            dpn << to_uid(*extensions).indent(2)
             dpn.OrgUid
             dpn.txtName(name)
             dpn.codeDatum('WGE')