+++ lib/aixm/feature/navigational_aid/marker.rb
@@ -11,16 +11,28 @@ module AIXM
 
         public_class_method :new
 
+        # TODO: Marker require an associated ILS
+#        def initialize(*args)
+#          fail(NotImplementedError, "Marker are not fully implemented yet")
+#        end
+
         ##
-        # Render AIXM
+        # Render UID markup
+        def to_uid(*extensions)
+          builder = Builder::XmlMarkup.new(indent: 2)
+          builder.MkrUid({ newEntity: (true if extensions >> :ofm) }.compact) do |mkruid|
+            mkruid.codeId(id)
+            mkruid.geoLat(xy.lat(format_for(*extensions)))
+            mkruid.geoLong(xy.long(format_for(*extensions)))
+          end
+        end
+
+        ##
+        # Render AIXM markup
         def to_xml(*extensions)
           builder = to_builder(*extensions)
           builder.Mkr do |mkr|
-            mkr.MkrUid({ newEntity: (true if extensions >> :ofm) }.compact) do |mkruid|
-              mkruid.codeId(id)
-              mkruid.geoLat(xy.lat(format_for(*extensions)))
-              mkruid.geoLong(xy.long(format_for(*extensions)))
-            end
+            mkr << to_uid(*extensions).indent(2)
             mkr.OrgUid
             mkr.valFreq(75)
             mkr.uomFreq('MHZ')